USE_KERNEL implement

+++ utils/dce_build.sh
@@ -1,5 +1,16 @@
 #!/bin/bash
 
+USE_KERNEL=NO
+args=("$@")
+NB=$#
+for (( i=0;i<$NB;i++)); do
+    if [ ${args[${i}]} = '-k' ]
+    then
+       USE_KERNEL=YES
+       WGET=wget
+    fi
+done
+
 cd `dirname $BASH_SOURCE`/../..
 
 # build quagga
@@ -73,34 +84,43 @@ patch -p1 < ../ns-3-dce-patches/120406-dce-quagga-support.patch
 cd ..
 
 # mod ns-3-linux (FIXME)
-hg clone http://202.249.37.8/ical/ns-3-linux-patches/
-cd ns-3-linux
-patch -p1 < ../ns-3-linux-patches/120406-linux-quagga-support.patch
-make clean
-#sed "s/uname -p/uname -m/" processor.mk >a
-#mv a processor.mk
-#sed 's/\$@/\$@\//g' Makefile.print >a
-#mv a Makefile.print
-rm -f config
-make config
+if [ "YES" == "$USE_KERNEL" ]
+then
 
-#sed "s/CONFIG_IPV6=m/CONFIG_IPV6=y/" config >a
-#mv a config
-#sed "s/case CAP_NET_RAW: return 1;/case CAP_NET_RAW: \n  case CAP_NET_BIND_SERVICE:\n  case CAP_NET_ADMIN: \n  return 1;/" sim/security.c > a
-#mv a sim/security.c
-#sed "s/msg->msg_iov = kernel_iov;/struct cmsghdr *user_cmsgh = msg->msg_control;\n  size_t user_cmsghlen = msg->msg_controllen;\n msg->msg_iov = kernel_iov;/" sim/sim-socket.c > a
-#sed "s/msg->msg_iov = user_iov;/msg->msg_iov = user_iov;\n  msg->msg_control = user_cmsgh; \n  msg->msg_controllen = user_cmsghlen - msg->msg_controllen;/" a >b
-#sed "s/size += msg->msg_iov->iov_len;/size += msg->msg_iov[i].iov_len;/" b > c
-#mv c sim/sim-socket.c
-
-make
-cd ..
+     hg clone http://202.249.37.8/ical/ns-3-linux-patches/
+     cd ns-3-linux
+     patch -p1 < ../ns-3-linux-patches/120406-linux-quagga-support.patch
+     make clean
+     #sed "s/uname -p/uname -m/" processor.mk >a
+     #mv a processor.mk
+     #sed 's/\$@/\$@\//g' Makefile.print >a
+     #mv a Makefile.print
+     rm -f config
+     make config
+     
+     #sed "s/CONFIG_IPV6=m/CONFIG_IPV6=y/" config >a
+     #mv a config
+     #sed "s/case CAP_NET_RAW: return 1;/case CAP_NET_RAW: \n  case CAP_NET_BIND_SERVICE:\n  case CAP_NET_ADMIN: \n  return 1;/" sim/security.c > a
+     #mv a sim/security.c
+     #sed "s/msg->msg_iov = kernel_iov;/struct cmsghdr *user_cmsgh = msg->msg_control;\n  size_t user_cmsghlen = msg->msg_controllen;\n msg->msg_iov = kernel_iov;/" sim/sim-socket.c > a
+     #sed "s/msg->msg_iov = user_iov;/msg->msg_iov = user_iov;\n  msg->msg_control = user_cmsgh; \n  msg->msg_controllen = user_cmsghlen - msg->msg_controllen;/" a >b
+     #sed "s/size += msg->msg_iov->iov_len;/size += msg->msg_iov[i].iov_len;/" b > c
+     #mv c sim/sim-socket.c
+     
+     make
+     cd ..
+fi
 
 # build ns-3-dce-quagga
+if [ "YES" == "$USE_KERNEL" ]
+then
+    WAF_KERNEL=--enable-kernel-stack=`pwd`/../ns-3-linux
+fi
+
 cd ns-3-dce-quagga
 . ../ns-3-dce/utils/setenv.sh
 cd ../ns-3-dce-quagga
-./waf configure --prefix=`pwd`/../build --verbose --enable-kernel-stack=`pwd`/../ns-3-linux
+./waf configure --prefix=`pwd`/../build --verbose $WAF_KERNEL
 ./waf
 ./waf install
 echo Launch NS3QUAGGATEST-DCE
